+++ Tests/Set-PASSafe.Tests.ps1
@@ -72,14 +72,14 @@ Describe $($PSCommandPath -Replace '.Tests.ps1') {
 					'SafeName' = 'SomeName'
 
 				}
-
+				Mock Get-PASSafe -MockWith {}
 				$response = $InputObj | Set-PASSafe -NumberOfDaysRetention 1 -ManagingCPM SomeCPM -NewSafeName SomeNewName -UseGen1API
 
 			}
 
 			It 'sends request' {
 
-				Assert-MockCalled Invoke-PASRestMethod -Times 1 -Exactly -Scope It
+				Assert-MockCalled Invoke-PASRestMethod -Scope It
 
 			}
 
@@ -89,13 +89,13 @@ Describe $($PSCommandPath -Replace '.Tests.ps1') {
 
 					$URI -eq "$($Script:psPASSession.BaseURI)/WebServices/PIMServices.svc/Safes/SomeName"
 
-				} -Times 1 -Exactly -Scope It
+				} -Times 1 -Scope It
 
 			}
 
 			It 'uses expected method' {
 
-				Assert-MockCalled Invoke-PASRestMethod -ParameterFilter { $Method -match 'PUT' } -Times 1 -Exactly -Scope It
+				Assert-MockCalled Invoke-PASRestMethod -ParameterFilter { $Method -match 'PUT' } -Times 1 -Scope It
 
 			}
 
@@ -107,7 +107,7 @@ Describe $($PSCommandPath -Replace '.Tests.ps1') {
 
 					($Script:RequestBody.safe) -ne $null
 
-				} -Times 1 -Exactly -Scope It
+				} -Scope It
 
 			}
 
@@ -125,7 +125,9 @@ Describe $($PSCommandPath -Replace '.Tests.ps1') {
 				Mock Invoke-PASRestMethod -MockWith {
 					[PSCustomObject]@{'Prop1' = 'Val1'; 'Prop2' = 'Val2' }
 				}
-
+				Mock Get-PASSafe -MockWith {
+					[PSCustomObject]@{'UpdateSafeResult' = [PSCustomObject]@{'description' = 'Val1'; 'location' = 'Val2' } }
+				}
 				$InputObj = [pscustomobject]@{
 					'SafeName' = 'SomeName'
 
@@ -137,7 +139,7 @@ Describe $($PSCommandPath -Replace '.Tests.ps1') {
 
 			It 'sends request' {
 
-				Assert-MockCalled Invoke-PASRestMethod -Times 1 -Exactly -Scope It
+				Assert-MockCalled Invoke-PASRestMethod -Scope It
 
 			}
 
@@ -153,7 +155,7 @@ Describe $($PSCommandPath -Replace '.Tests.ps1') {
 
 			It 'uses expected method' {
 
-				Assert-MockCalled Invoke-PASRestMethod -ParameterFilter { $Method -match 'PUT' } -Times 1 -Exactly -Scope It
+				Assert-MockCalled Invoke-PASRestMethod -ParameterFilter { $Method -match 'PUT' } -Times 1 -Scope It
 
 			}
 
@@ -165,13 +167,13 @@ Describe $($PSCommandPath -Replace '.Tests.ps1') {
 
 					$Script:RequestBody -ne $null
 
-				} -Times 1 -Exactly -Scope It
+				} -Times 1 -Scope It
 
 			}
 
 			It 'has a request body with expected number of properties' {
 
-				($Script:RequestBody | Get-Member -MemberType NoteProperty).length | Should -Be 3
+				($Script:RequestBody | Get-Member -MemberType NoteProperty).length | Should -Be 5
 
 			}
 
@@ -226,7 +228,7 @@ Describe $($PSCommandPath -Replace '.Tests.ps1') {
 					'SafeName' = 'SomeName'
 
 				}
-
+				Mock Get-PASSafe -MockWith {}
 				$response = $InputObj | Set-PASSafe -NumberOfDaysRetention 1 -ManagingCPM SomeCPM -NewSafeName SomeNewName
 
 			}