+++ psPAS/Functions/AccountACL/Add-PASAccountACL.ps1
@@ -62,7 +62,9 @@ function Add-PASAccountACL {
 
 	)
 
-	BEGIN { }#begin
+	BEGIN {
+		Assert-VersionRequirement -SelfHosted
+	}#begin
 
 	PROCESS {