+++ psPAS/Functions/Authentication/Set-PASOpenIDConnectProvider.ps1
@@ -42,7 +42,7 @@ Function Set-PASOpenIDConnectProvider {
 		[string]$description,
 
 		[parameter(
-			Mandatory = $true,
+			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $true
 		)]
 		[ValidateNotNullOrEmpty()]
@@ -56,7 +56,7 @@ Function Set-PASOpenIDConnectProvider {
 		[string]$jwkSet,
 
 		[parameter(
-			Mandatory = $true,
+			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $true
 		)]
 		[ValidateLength(1, 100)]
@@ -71,7 +71,7 @@ Function Set-PASOpenIDConnectProvider {
 		[securestring]$clientSecret,
 
 		[parameter(
-			Mandatory = $true,
+			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $true
 		)]
 		[ValidateSet('Basic', 'Post')]
@@ -110,6 +110,11 @@ Function Set-PASOpenIDConnectProvider {
 			#Include decoded clientSecret in request
 			$boundParameters['clientSecret'] = $(ConvertTo-InsecureString -SecureString $clientSecret)
 
+		}
+
+		$OIDCProvider = Get-PASOpenIDConnectProvider -id $id
+		if ($null -ne $OIDCProvider) {
+			Format-PutRequestObject -InputObject $OIDCProvider -boundParameters $BoundParameters -ParametersToRemove id
 		}
 
 		#Create body of request