+++ psPAS/Functions/EventSecurity/Set-PASPTARule.ps1
@@ -10,40 +10,40 @@ Function Set-PASPTARule {
 		[string]$id,
 
 		[parameter(
-			Mandatory = $true,
+			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $true
 		)]
 		[ValidateSet('SSH', 'WINDOWS', 'SCP', 'KEYSTROKES', 'SQL')]
 		[string]$category,
 
 		[parameter(
-			Mandatory = $true,
+			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $true
 		)]
 		[string]$regex,
 
 		[parameter(
-			Mandatory = $true,
+			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $true
 		)]
 		[ValidateRange(1, 100)]
 		[int]$score,
 
 		[parameter(
-			Mandatory = $true,
+			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $true
 		)]
 		[string]$description,
 
 		[parameter(
-			Mandatory = $true,
+			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $true
 		)]
 		[ValidateSet('NONE', 'TERMINATE', 'SUSPEND')]
 		[string]$response,
 
 		[parameter(
-			Mandatory = $true,
+			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $true
 		)]
 		[boolean]$active,
@@ -90,6 +90,11 @@ Function Set-PASPTARule {
 		#Get all parameters that will be sent in the request
 		$boundParameters = $PSBoundParameters | Get-PASParameter
 
+		$PTARule = (Get-PASPTARule).Where({ $PSitem.ID -eq $id })
+		if ($null -ne $PTARule) {
+			Format-PutRequestObject -InputObject $PTARule -boundParameters $BoundParameters -ParametersToRemove mode, list
+		}
+
 		#Create URL for Request
 		$URI = "$($psPASSession.BaseURI)/API/pta/API/Settings/RiskyActivity/"
 
@@ -114,19 +119,47 @@ Function Set-PASPTARule {
 				if (-not($boundParameters['scope'].ContainsKey($scopeItem))) {
 					$boundParameters['scope'].Add($scopeItem, @{})
 				}
-				#translate paramer names into request property name
+				#translate parameter names into request property name
 				#* Return only last 4 characters of parametername in lowercase
 				#*vaultUsersMode & machinesMode translate to "mode"
 				#*vaultUsersList & machinesList translate to "list"
 				$property = ($PSItem).Substring(($PSItem).length - 4, 4).ToLower()
 
-				#Add scope parameter vaultUsers & machines request values to boundParameters
-				$boundParameters['scope'][$scopeItem][$property] = $PSBoundParameters[$PSItem]
+				if ($null -ne $PSBoundParameters[$PSItem]) {
+					#Add scope parameter vaultUsers & machines request values to boundParameters
+					$boundParameters['scope'][$scopeItem][$property] = $PSBoundParameters[$PSItem]
+				} else {
+					#clear vaultUsers/machine scope if null value specified
+					$boundParameters['scope'][$scopeItem] = $null
+				}
+
 
 			}
 
 		}
 
+		switch ($boundParameters) {
+			{ $PSItem.keys -contains 'scope' } {
+				#BoundParameteters Contains Scope Key
+				switch ($boundParameters['scope']) {
+					{ $PSItem.ContainsKey('vaultUsers') -and (-not($PSItem.ContainsKey('machines'))) } {
+						#vaultUser Scope is being updated - copy existing machines scope
+						$boundParameters['scope']['machines'] = $PTARule.scope.machines
+					}
+
+					{ $PSItem.ContainsKey('machines') -and (-not($PSItem.ContainsKey('vaultUsers'))) } {
+						#machines Scope is being updated - copy existing vaultUsers scope
+						$boundParameters['scope']['vaultUsers'] = $PTARule.scope.vaultUsers
+					}
+				}
+
+			}
+			default {
+				#No updated scope specified - Use existing scope copied from rule
+				'third'
+				$boundParameters['scope'] = $PTARule.scope
+			}
+		}
 		#Create body of request
 		#* Ensure JSON Depth of 3 to capture correct scope format
 		$Body = $boundParameters | Get-PASParameter -ParametersToRemove $scopeParams | ConvertTo-Json -Depth 3