+++ psPAS/Functions/Safes/Set-PASSafe.ps1
@@ -53,7 +53,7 @@ function Set-PASSafe {
 		[string]$ManagingCPM,
 
 		[Parameter(
-			Mandatory = $true,
+			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $true,
 			ParameterSetName = 'Gen2-NumberOfVersionsRetention'
 		)]
@@ -66,7 +66,7 @@ function Set-PASSafe {
 		[int]$NumberOfVersionsRetention,
 
 		[Parameter(
-			Mandatory = $true,
+			Mandatory = $false,
 			ValueFromPipelinebyPropertyName = $true,
 			ParameterSetName = 'Gen2-NumberOfDaysRetention'
 		)]
@@ -100,10 +100,22 @@ function Set-PASSafe {
 
 		$BoundParameters = $PSBoundParameters | Get-PASParameter -ParametersToRemove NewSafeName
 
-		if ($PSBoundParameters.ContainsKey('NewSafeName')) {
+		$SafeObject = Get-PASSafe -SafeName $SafeName
+		if ($null -ne $SafeObject) {
+			Format-PutRequestObject -InputObject $SafeObject -boundParameters $BoundParameters -ParametersToKeep ManagingCPM, location, Description,
+			NumberOfVersionsRetention, NumberOfDaysRetention
+		}
 
-			$BoundParameters['SafeName'] = $PSBoundParameters['NewSafeName']
-
+		switch ($PSBoundParameters.Keys) {
+			'NewSafeName' {
+				$BoundParameters['SafeName'] = $PSBoundParameters['NewSafeName']
+			}
+			'NumberOfDaysRetention' {
+				$BoundParameters.Remove('NumberOfVersionsRetention')
+			}
+			'NumberOfVersionsRetention' {
+				$BoundParameters.Remove('NumberOfDaysRetention')
+			}
 		}
 
 		switch ($PSCmdlet.ParameterSetName) {