+++ psPAS/Functions/ServerWebServices/Get-PASSafeShareLogo.ps1
@@ -10,7 +10,7 @@ function Get-PASSafeShareLogo {
 	)
 
 	BEGIN {
-		#!Depreciated above 13.2
+		#!Depracated above 13.2
 		Assert-VersionRequirement -MaximumVersion 13.2
 	}#begin